+++ src/Display/Screens/Components/Date.h
@@ -10,12 +10,12 @@ class Date {
 
 		protected:
 
-			const uint POS_Y = 18;
+			const uint8_t POS_Y = 18;
 
-			uint _prevDay = 99;
+			uint8_t _prevDay = 99;
 
 			void _renderDate(RTC_Date currentDate);
 			void _renderDayInWeek(RTC_Date currentDate);
-			void _weekday(char *dayInWeekStr, int year, int month, int day);
+			void _weekday(char *dayInWeekStr, uint year, uint8_t month, uint8_t day);
 
 };
