+++ src/Display/Screens/Components/InfoPanelComponents/Battery.h
@@ -12,12 +12,12 @@ class Battery {
 
 	protected:
 
-		int _prevCapacity = 101;
+		uint8_t _prevCapacity = 101;
 		bool _prevChargingState = false;
 
 		const uint POS_X = 15;
 		const uint POS_Y = 140;
 
-		void _renderBatteryIcon(int capacity);
+		void _renderBatteryIcon(uint8_t capacity);
 		void _renderChargingIcon();
 };
