+++ src/Display/Screens/Components/InfoPanelComponents/StepCounter.h
@@ -15,9 +15,9 @@ class StepCounter {
 
 	protected:
 
-		const uint POS_X = 20;
-		const uint POS_Y = 170;
-		uint _prevStepCount = 0;
+		const uint8_t POS_X = 20;
+		const uint8_t POS_Y = 170;
+		uint8_t _prevStepCount = 0;
 		uint _iconRendered = false;
 
 		void _clearDisplay();
